CB-9901 Update load balancer when instances change
Adds new logic to AwsDownscaleService and AwsUpscaleService. The new logic will
update the load balancer's routing rules to either remove any instances that
are being deleted from the instance groups the load balancer routes traffic to,
or add any new instances in those groups to the load balancer's targets.

Tested with unit tests, and by terminating the Knox gateway node and initating
a repair to verify the load balancer was correctly updated.