CB-10162 Add log message every time the FreeIPA status is checked
During ENGESC-5655, we had a hard time debugging the freeipa status. Currently we only log freeipa status when there is a transition (eg. from healthy to unreachable) and it was determined that adding a log message for both freeipa status and instance status during health check would be helpful. Now we added logging for every health check.